+++ wrappers/ReportsWrapper.jsx
@@ -1,7 +1,0 @@
-"use client";
-
-import { ReportsProvider } from "@/context/ReportsContext";
-
-export default function ReportsWrapper({ children }) {
-  return <ReportsProvider>{children}</ReportsProvider>;
-}